+++ packages/gatsby-cli/src/reporter/redux/types.ts
@@ -43,7 +43,7 @@ interface ILog {
   group: string | undefined
   code: string | undefined
   type: string | undefined
-  category?: ErrorCategory
+  category?: keyof typeof ErrorCategory
   filePath: string | undefined
   location: IStructuredError["location"] | undefined
   docsUrl: string | undefined