+++ packages/gatsby-plugin-utils/src/validate.ts
@@ -1,7 +1,8 @@
-import { BaseValidationOptions } from "joi"
-import { ObjectSchema } from "./utils/plugin-options-schema-joi-type"
+import { ValidationOptions } from "joi"
+import { ObjectSchema } from "./joi"
+import { IPluginInfoOptions } from "gatsby"
 
-const validationOptions: BaseValidationOptions = {
+const validationOptions: ValidationOptions = {
   // Show all errors at once, rather than only the first one every time
   abortEarly: false,
   cache: true,
@@ -11,19 +12,19 @@ interface IOptions {
   validateExternalRules?: boolean
 }
 
-export async function validateOptionsSchema<PluginOptions = object>(
+export async function validateOptionsSchema(
   pluginSchema: ObjectSchema,
-  pluginOptions: PluginOptions,
-  options: IOptions = {}
-): Promise<PluginOptions> {
-  if (options.validateExternalRules === false) {
-    const result = pluginSchema.validate(pluginOptions, {
-      ...validationOptions,
-      externals: false,
-    })
-    if (result.error) throw result.error
-    return result.value
+  pluginOptions: IPluginInfoOptions,
+  options: IOptions = {
+    validateExternalRules: true,
   }
+): Promise<IPluginInfoOptions> {
+  const { validateExternalRules } = options
 
-  return pluginSchema.validateAsync(pluginOptions, validationOptions)
+  const value = await pluginSchema.validateAsync(pluginOptions, {
+    ...validationOptions,
+    externals: validateExternalRules,
+  })
+
+  return value
 }